[Refactor] inital_state 수정
